refactor: Rename Dagster package in line with the use case
