+++ deploy/dagster/Dockerfile.salary_prediction
@@ -54,4 +54,4 @@ EXPOSE 4000
 
 ENTRYPOINT ["uv", "run"]
 # CMD allows this to be overridden from run launchers or executors to execute runs and steps
-CMD ["dagster", "code-server", "start", "-h", "0.0.0.0", "-p", "4000", "-m", "income_prediction"]
+CMD ["dagster", "code-server", "start", "-h", "0.0.0.0", "-p", "4000", "-m", "salary_prediction"]
