+++ src/salary_prediction/constants.py
