+++ src/salary_prediction/jobs.py
