+++ src/salary_prediction/types.py
